+++ examples/bunnymark/example.rs
@@ -174,13 +174,14 @@ impl Example {
         let mut command_encoder = context.create_command_encoder(gpu::CommandEncoderDesc {
             name: "init",
             buffer_count: 1,
+            queue: gpu::QueueType::Main,
         });
         command_encoder.start();
         command_encoder.init_texture(texture);
         if let mut transfer = command_encoder.transfer("init texture") {
             transfer.copy_buffer_to_texture(upload_buffer.into(), 4, texture.into(), extent);
         }
-        let sync_point = context.submit(&mut command_encoder);
+        let sync_point = context.submit(&mut command_encoder, &[]);
         let _ = context.wait_for(&sync_point, !0);
 
         context.destroy_command_encoder(&mut command_encoder);